Began refactoring base async matcher classes to allow for matching an event dispatch x number of times.
Doing this in order to allow for a test looking for an event with specific data to timeout even if the right event with wrong data has been dispatched. I still want to allow for this case to straight up fail, and by asserting on an event being dispatched exactly once, and with specific data, would allow for that.